requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
apify_fingerprint_datapoints==0.12.0
APScheduler==3.11.2
attrs==26.1.0
audioop-lts==0.2.2
blinker==1.9.0
browserforge==1.2.4
camoufox==0.4.11
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
Cython==3.2.4
filelock==3.28.0
Flask==3.1.3
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
geoip2==5.2.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
greenlet==3.4.0
gspread==6.2.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.3
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
language-tags==1.2.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
maxminddb==3.1.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
opencv-python==4.13.0.92
orjson==3.11.8
packaging==26.1
patchright==1.58.2
pillow==12.2.0
platformdirs==4.9.6
playwright==1.58.0
playwright-recaptcha==0.5.1
playwright-stealth==2.0.3
polars==1.40.0
polars-runtime-32==1.40.0
propcache==0.4.1
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydub==0.25.1
pyee==13.0.1
Pygments==2.20.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
recognizer==1.4
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
scipy==1.17.1
screeninfo==0.8.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
SpeechRecognition==3.10.4
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==8.3.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing_extensions==4.15.0
tzlocal==5.3.1
ua-parser==1.0.2
ua-parser-builtins==202603
ultralytics==8.4.38
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
